DataViews: Bootstrap Quick Edit
