+++ packages/edit-site/src/components/post-edit/style.scss
@@ -1,0 +1,3 @@
+.edit-site-post-edit {
+	padding: $grid-unit-30;
+}
